requirements.txt
streamlit
opencv-python-headless
torch
torchvision
transformers
ultralytics
Pillow
matplotlib
psycopg2-binary